requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
multidict==6.0.5
openai==0.28.0
packaging==24.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
